# Improve Enum.ToString perf for [Flags] enums